// appth/src/app/models/HojadeVidaBusquedaVO.ts
export class HojadeVidaBusquedaVO {
  id?: number;
  primernombre?: string;
  segundonombre?: string;
  primerapellido?: string;
  segundoapellido?: string;
  tipodocumento?: string;
  estadoCivil?: string;
  fechanacimiento?: Date;
  telefonocasa?: string;
  numerodocumento?: string;
  sexo?: string;
  correoelectronico?: string;
  telefonomovil?: string;
  profesion?: number;
  cargo?: string;
  area?: string;
  constructor() {}
}