// appth/src/app/models/contratoVO.ts
export class ContratoVO {
    contrato: number;
    documentoidentificacion: String;
    numeroContrato: String;
    cargo: number;
    area: number;
    arl: number;
    fondoPension: number;
    sede: number;
    tipoContrato: number;
    estadoContrato: number;
    eps: number;
    objetivo: String;
    fechaInicio: String;
    fechaFin: String;
    salario: number;
    tallaCamisa: String;
    tallaPantalon: String;
    tallaCalzado: number;
    idDotacion: number;
    doccontrato: String;
}