// appth/src/app/models/EditarContratoVO.ts
import { HojadeVidaBusquedaVO } from './HojadeVidaBusquedaVO';
export class EditarContratoVO {
    contrato: number;
    hojaVida: number;
    numeroContrato: String;
    cargo: number;
    area: number;
    arl: number;
    fondoPension: number;
    sede: number;
    tipoContrato: number;
    estadoContrato: number;
    eps: number;
    objetivo: String;
    fechaInicio: Date;
    fechaFin: Date;
    salario: number;
    tallaCamisa: String;
    tallaPantalon: String;
    tallaCalzado: number;
    doccontrato: String;
    persona: HojadeVidaBusquedaVO;
}